Update to add createFetch
